[MOV] sale_management: split sale_order.py
