+++ addons/sale_management/models/__init__.py
@@ -4,4 +4,6 @@ from . import digest
 from . import res_company
 from . import res_config_settings
 from . import sale_order
+from . import sale_order_line
+from . import sale_order_option
 from . import sale_order_template
