+++ addons/sale_management/models/sale_order_line.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+
+
+class SaleOrderLine(models.Model):
+    _inherit = "sale.order.line"
+    _description = "Sales Order Line"
+
+    sale_order_option_ids = fields.One2many('sale.order.option', 'line_id', 'Optional Products Lines')
+
+    @api.depends('product_id')
+    def _compute_name(self):
+        # Take the description on the order template if the product is present in it
+        super()._compute_name()
+        for line in self:
+            if line.product_id and line.order_id.sale_order_template_id:
+                for template_line in line.order_id.sale_order_template_id.sale_order_template_line_ids:
+                    if line.product_id == template_line.product_id:
+                        line.name = template_line.with_context(lang=line.order_id.partner_id.lang).name + line._get_sale_order_line_multiline_description_variants()
+                        break
